+++ examples/src/main/java/org/eqasim/examples/corsica_vdf/RunCorsicaVDFSimulation.java
@@ -2,12 +2,14 @@ package org.eqasim.examples.corsica_vdf;
 
 import java.net.URL;
 
+import org.eqasim.core.components.traffic.EqasimTrafficQSimModule;
 import org.eqasim.core.simulation.analysis.EqasimAnalysisModule;
 import org.eqasim.core.simulation.mode_choice.EqasimModeChoiceModule;
 import org.eqasim.ile_de_france.IDFConfigurator;
 import org.eqasim.ile_de_france.mode_choice.IDFModeChoiceModule;
 import org.eqasim.vdf.VDFConfigGroup;
 import org.eqasim.vdf.VDFModule;
+import org.eqasim.vdf.VDFQSimModule;
 import org.matsim.api.core.v01.Scenario;
 import org.matsim.core.config.CommandLine;
 import org.matsim.core.config.CommandLine.ConfigurationException;
@@ -29,6 +31,8 @@ public class RunCorsicaVDFSimulation {
 				.build();
 
 		IDFConfigurator configurator = new IDFConfigurator();
+		configurator.getQSimModules().removeIf(m -> m instanceof EqasimTrafficQSimModule);
+
 		URL configUrl = Resources.getResource("corsica/corsica_config.xml");
 		Config config = ConfigUtils.loadConfig(configUrl, configurator.getConfigGroups());
 
@@ -46,6 +50,7 @@ public class RunCorsicaVDFSimulation {
 		controller.addOverridingModule(new IDFModeChoiceModule(cmd));
 
 		controller.addOverridingModule(new VDFModule());
+		controller.addOverridingQSimModule(new VDFQSimModule());
 
 		controller.run();
 	}